--- lib/internal/cluster.js
@@ -1,4 +1,0 @@
-'use strict';
-
-// Used in tests.
-exports.handles = {};
