+++ node.gyp
@@ -76,7 +76,12 @@
       'lib/zlib.js',
       'lib/internal/buffer.js',
       'lib/internal/child_process.js',
-      'lib/internal/cluster.js',
+      'lib/internal/cluster/child.js',
+      'lib/internal/cluster/master.js',
+      'lib/internal/cluster/round_robin_handle.js',
+      'lib/internal/cluster/shared_handle.js',
+      'lib/internal/cluster/utils.js',
+      'lib/internal/cluster/worker.js',
       'lib/internal/freelist.js',
       'lib/internal/fs.js',
       'lib/internal/linkedlist.js',
